rewrite the core AST node classes to simplify and uniformize
and make subclassing easier. Also, they are now produced by a factory function, to allow changes of the abstract base class `Node` to be inherited by its subclasses.

Encoding operators by a single character stored in `Operator.operator` is much more succinct and easier to check, than creating a different class for each one. The latter increases inhomogeneity of AST classes, making reading them, reusing them, writing functions that manipulate them, and subclassing convoluted, tedious, and error-prone. Alan Perlis applies again: simpler classes can be manipulated by more functions with less difficulty.

Lexical substitution of operators must be performed by the lexer --- *not* by the AST classes. So token rules like eventually, that recognizes both `<>` and `F`, must be implemented with token methods in the lexer, that for example recognizes both `<>` and `F`, but returns only `F`. The lexer is the appropriate tool designed for this job.

Note that using methods instead of plain regular expressions for some token rules affects the precedence, and so they need to be re-ordered, because `ply` maintains the order of token rule functions the same as they appear in the lexer source code. For example "EVENTUALLY" would shadow "False" (stealing the `F` from it).
Add some unit tests to guard against such cases in the future.

The operator map is useful for translations and a node factory is used, so that other languages can change the operator map for the `Node` abstract base class and have that change inherited also by the other nodes.

The operator `==` is removed from the lexer, because it can only cause sorrow.

The _eval methods are not used anywhere, because flattening to python and compiling to `code` objects is faster. So they are removed.

The maps[lang][self.op] are not replaced by a change of `opmap` that happens in `spec.translation` by writing new node class factories.