load `json` in `slugs` interface and update its tests
rm `jtlv` test on var attributes of `nx.Digraph` strategy, because they are not used in `synth.strategy_to_mealy` and thus may be deprecated and removed in the future.

The `GRSpec` itself is used already, of course this prevents the interface from loading a strategy w/o knowing the spec. But that is not the currently the use case, and modifying to that is straightforward, if it is ever needed.