CONTRIB: fmu Makefile: move extern dirs to vars; only use -lrt on Linux
ECOSDIR and FMIinclude are defined near the top of fmu/Makefile to
point to corresponding directories under extern/ in the TuLiP source
tree.

The `-lrt` switch is only used if building on Linux.  To achieve this,
we use `uname` as in the Makefile for ECOS.

The man page for clock_gettime() indicates that linking `rt` is only
required for earlier versions of glibc.