EXTERN: DOC, and return error if checksums do not match for ECOS
The intent is that get.sh should only return with success if it
actually had success, which is important if get.sh is used by something
else automatically.

The README file together with error messages or comments in get.sh
should be enough for users with little or no experience with shell
scripting.