CONTRIB: Output most temporary files under build/
However, some of the generated files are placed under the sources/
and include/ directory, rendering difficult the placement of all
parts of the FMU under build/.  While there are some manners to
still achieve this, e.g., using links, the most portable solution
would be copying the entire directories for placement under build/.

However, because it remains to integrate code under contrib/fmu into
Tulip, we may expect this build procedure to change anyway.  Further,
a user who wants to examine intermediate results of the build process
could be confused by the file duplicates.  Thus I recommend that those
files appearing outside of the build/ directory are targeted for
removal by

  make clean

as achieved in this commit.