Next version will be 1.2.0, instead of 1.2a
Henceforth we shall use all numbers, so 1.2.0, 1.2.1, ...
instead of 1.2a, 1.2b, ...

The bump commit in anticipation of the next release was
68369138c3c4c1d8be0f490f4b23119de96106fc
and this commit should be regarded as overriding that one.

The change to versions being only numbers is motivated by
PEP 440 (https://www.python.org/dev/peps/pep-0440/)
and, generally, to facilitate machine-reading of versions.