add attr: `type` to AST nodes
to distinguish them w/o the need to call `isinstance` or `type(node).__name__`.
This is a robust implementation that allows manipulating the AST w/o importing or even knowing where the node classes were created.

Using `isinstance` can break if the factory is called somewhere else. Otherwise we have to always import from where the nodes were created, possibly causing circular imports and certainly poking into external modules or packages.

Using the class name can break as soon as a subclass with a different name is created.

For distinguishing between `Terminal` and `Operator`, prefer the minimal check: use `hasattr` for `value` and `operator` (or `operands`), respectively.

If it quacks like a duck, then it probably is a duck.