Add 'wring' as option of translate() and translate_ast()
Thus the syntax of Lily is supported again, after the revision of
tulip.spec.ast and the creation of tulip.spec.translation.

GRSpec has been explicitly added as a possible type for the `formula`
parameter of interfaces.lily.synthesize(). If `formula` is an instance
of GRSpec, then translate() is invoked; else, parse() and then
translate_ast(). While GRSpec inherits from LTL and thus may be treated
as such in lily.synthesize(), we are motivated to begin using translate()
where possible in anticipation of it being a general purpose routine that
treats more than GR(1) formulae.

Note that the assumption of GRSpec by translate() is reflected by the
subsequent target-specific functions, e.g., to_gr1c(), which rely on this
special structure. As is, _to_wring copes by creating a single LTL formula
string (in Wring syntax) from the usual GR(1)-oriented pieces.