BUGFIX: Rollback edges_iter in FiniteTransitionSystem.__str__
The temporary introduction of natsort in
9bf4fb42583ac9e7c77449e62cae2f12f1858e92, which has since been removed
in 09f4d963d3eafc9cba84ede5a5220d23fcc3b757, included modification of the
__str__() of FiniteTransitionSystem to call pprint.pformat() on
self.edges_iter(data=True) rather than self.transitions(data=True). But
this fails because the former is a generator. As an illustration,

>>> from pprint import pformat
>>> pformat(range(10))
'[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]'
>>> pformat(xrange(10))
'xrange(10)'

Apparently this bug was not noticed because it only manifested when natsort
was not used, yet natsort was enabled by default and it would use
self.edges(data=True) and return the list (not a generator).