Added functionality to discretize mode-only switched systems
Added AbstractModeOnlySwitched class which includes the progress map
for the system, and allows plotting of the transitions in the system.
The function discretize_modeonlyswitched acts as a wrapper function that
takes in the dynamics and discretizes it, creates an augmented finite
transition system and a progress group map for it.

WARNING: synthesis of Mode-Only Switched Systems is not working properly
yet. Using synthesize on these systems may cause an error.