DOC: Improve style of bibliography; generate it from bib.txt
In this changeset, the bibiography style has been improved, and the
format of the file in which the bibliography is stored has been simplified.

In summary, entries are made in doc/bib.txt, and `make html` causes the
script genbib.py to be called, whence the reST file bibliography.rst is
generated. I claim that the style is improved because the previous format
does not left-align references. Furthermore, the switch to bib generation
may ease later changing of the source file to use bibtex.

This changeset also provides new instructions in the Developer's Guide
for creating and citing references.