push slugs options lower, to `_call_slugs`
so that they reside at a single place, and be available also to whoever wants to import and call the wrapper `_call_slugs` independently. This is a valid use case when the "slugsin" (prefix syntax) input is generated independently, and not by using the compiler included in slugs. See also comments in c13f08daf86225797a1b1ccc39f8b47bc5962173.

Add option for telling `slugs` to dump a symbolic representation of the strategy to a DDDMP file. This is useful when an enumerated strategy becomes too large.