BUGFIX: in double_tank.py example, enclose "string literals"
Variables with arbitrary finite domains must have values enclosed in
double quotations. These values are referred to as constants,
literals, or string literals. Without double quotes, they are regarded
as variable names.

Note that there is a function for automatically adding quotes when
inference is possible. infer_constants() in tulip.spec.transformation
However, for the example it is better to explicitly illustrate the
TuLiP LTL syntax, rather than rely on an inference routine.