From dumpsmach.python_case(), include list of input variables
In the code generated by tulip.dumpsmach.python_case(), add a list
attribute `input_vars` to the class that provides the names of input
variables in the same order as parameters to the move() method.

A motivating use-case is automatically generated LTL formulae in which
variable names are not fixed a priori. For other programs to
incorporate the output of dumpsmach.python_case(), the input variable
names or, for invocations using positional parameters, a count of them
must be known. E.g., a list of names facilitates the creation of a
dictionary `kwargs` then calls like `move(**kwargs)`.