dumpsmach.python_nx(), autocoding to Python and NetworkX
The generated code is independent of TuLiP (also known as "standalone").
It is in Python, and the only dependency not in the standard library
is `networkx`. Usage of the generated class is the same as from
dumpsmach.python_case(). The Mealy machine is represented as a
networkx.MultiDiGraph object, which is saved within the generated code
as compressed pickled data.

Optionally, compression may not be used, in which case the generated
code does not depend on zlib.

This work is motivated by an application where CPython fails to import
code generated by dumpsmach.python_case(), perhaps because the file is
too large (> 49 MB) or the definition of move() is too long (1152188 lines).
As a graph, the Mealy machine has 722 nodes and 104544 edges.
Exporting the same controller via dumpsmach.python_nx() with and
without compression results in files with sizes less than 3.2 MB and
47 MB, respectively, and both of which can be imported and have the
corresponding classes instantiated quickly.