List possible inputs in exception message from reaction()
The behavior depends on whether lazy=True when invoking
MealyMachine.reaction(). If lazy, then the suggested possibilities
only include assignments of values to the input variables used during
the attempted call to reaction(). If not lazy, then the possibilities
include valuations of all input variables.
Cf. issue #129 and pull request #130.

The exception message is as proposed by Kaushik Kannan (@iamkaushik5 on GitHub).