BUGFIX: Return None when request to synthesize() unrealizable
and a regression test for the same. The bug arose from two causes.
First, different solver interfaces return values of different types if
the specification is not realizable. Second, the function that
constructs a MealyMachine object from the returned value was placed
before any check about the type of that value.

Closes #120