Delete JTLV-based solver; create infrastructure to fetch it
The fetching script is extern/get-jtlv.sh. The comments therein
descibe usage and what it does. I have only tested it on Linux x86_64.
Note that it uses cURL, whereas extern/ecos/get.sh uses wget.

This changeset is nontrivial because we now manipulate package data
depending on the presence of an optional item. Messages emitted when
attempting to install or check dependencies of TuLiP should be
sufficienly descriptive to aid users who want the JTLV-based solver.
The code introduced here should be useful for similar cases of
predicated package data in the future.

Closes #123