Use JSON as default format in gr1c interface
tulip.interfaces.gr1c.load_mealy() now has an optional parameter
`fformat` to specify which output format from gr1c is to be used. It
can take values "json" and "xml". Default is "xml"; thus, any existing
code that used load_mealy() does not need to be changed.

N.B., the choice of "tulipxml" as a format name in load_mealy() is
motivated by it corresponding to the switch `-t tulip` of gr1c. It
should not be shortened to "xml" because gr1c may eventually have the
capability to output into an xml schema other than tulipcon.