REL: rm custom check for `cvxopt` from `setup.py`
because only `cvxopt.glpk` is intended to be used, and that is checked separately by the function `check_glpk`.
In any case, `cvxopt` is listed in `extras_require` and `requirements.txt`.

The main usage of `cvxopt` is in package `polytope`.
In `tulip`, `cvxopt` is explicitly imported only in `tulip.abstract.find_controller`.

In the future, the Python solver of `cvxopt` can be replaced with `scipy.optimize.linprog`.