REL: mv check for `gr1c` from `setup.py` to `tulip.interfaces.gr1c`
to localize checking for `gr1c` to its Python interface.
This will raise an exception when invoked, which is more likely to be noticed.
It lets `tulip` be installed in absence of `gr1c`, so that functionality other than synthesis be used.

In the future, the interfaces to the Python solvers in packages `gr1py` and `omega` may be used as default solvers for games.