DEP: rm infrastructure for checking dependencies
most of it was necessary when more heterogeneous and non-Python dependencies where required.
Since then, `pip` has improved noticeably, so it can serve this purpose for the Python dependencies.

Each module in `tulip.interfaces` is responsible for checking the presence of the tool it communicates with.
This arrangement is local and simpler.