DOC: Delete reference to defunct example
examples/gridworlds/smallsurvey.py

Similar code that demonstrates creation of an initial partition and
refinement according to dynamics is given in the section on gridworlds
in the TuLiP User's Guide.

smallsurvey.py also included synthesis and simulated execution, which
might be worth providing again as part of an example eventually.