DOC: Use 's' instead of C{s} in GridWorld.loads docstring
for symbols `s` that can be used in gridworld description. The
motivation to change is that two use-cases are not addressed well with
C{s}: inclusion in the TuLiP User's Guide and reading the docstring
directly (i.e., without a markup renderer). The syntax `C{s}` is that
of Epydoc, but simply writing 's' should still be understood in the
Epydoc-generated API manual.