rm obsolete GridWorld.discreteTransitionSystem()
It was originally used for synthesis with NuSMV.  Consult commit
d8e3fa64275a93a4be88eca297725c537ed58333.

Later it was intended to be useful generally as providing a "discrete
transition system" based on an instance of PropPreservingPartition
without polytopes. Since the time of that implementation, we now have
FiniteTransitionSystem and other classes under tulip.transys that are
better for the same purpose.