DOC: rm suggestion to use Debian packages for SciPy etc.
At the time of writing, `pip` and `scipy` as available in the stable
Debian repositories are excessively old. E.g., `polytope` depends on a
more recent minimum version of `scipy`.

As such, users of TuLiP should not be advised to use Debian packages
in the installation page of the User's Guide.