API: revert to union over all horizon only when `use_all_horizon`
because allowing the controller to choose the horizon affects
the timing semantics of when the system observes changes
to discrete-valued variables of the environment.

In between observations (so within the horizon),
discrete-valued environment variables are
assumed to remain unchanged.

Thus, if the system picks the horizon, in effect is also picks the
sampling times. So, the assumptions that allow concretization of
a discrete controller are different for each value of `use_all_horizon`.

This change partially reverts:
    55a436712c6632cbdad13aed9271ea0e1e3366db
which appears to have modified the case `use_all_horizon`,
but also, in so doing, forgotten the case `not use_all_horizon`.

The resulting algorithm:

- for `not use_all_horizon` is the succession of polytopes in:
  https://github.com/tulip-control/tulip-control/blob/92f717f4b381714f8edcb659b27f3b9791a57904/tulip/abstract/feasible.py#L167
  (chain technically means a nested collection of sets,
   so something different)

- for `use_all_horizon` is the modified computation in:
  https://github.com/tulip-control/tulip-control/blob/55a436712c6632cbdad13aed9271ea0e1e3366db/tulip/abstract/feasible.py#L155
  which is the same with `not use_all_horizon` in terms of the
intermediate results, but the result is the union of these intermediate
results.
  In other words, this case is equivalent to calling `not
use_all_horizon` with N different horizons 1..N,
  and returning the union of the results.

  It is *different* than the attractor approximation in:
  https://github.com/tulip-control/tulip-control/blob/92f717f4b381714f8edcb659b27f3b9791a57904/tulip/abstract/feasible.py#L165