Update interface for slugs
The interface is now designed with respect to commit
ad0cf12c14131fc6a20fe29edfe04d9aefd7c6d4 of the slugs Git repository,
which is the current tip of the master branch. They do not provide
tags nor versions (nor releases), so we use commit hashes as references.