TEST: Bugfix test case of gr1c.check_realizable()
Commit 1965df31f7c3f50497972a7cdffc99f52449a603 introduced a bug
in one of the test cases for the function check_realizable of
tulip.interfaces.gr1c. The variables should only take values over
the Boolean domain, but `env_init` was assigned to have 'y = 0'
and 'y = 5'.

The forthcoming version of gr1c aborts on that specification, so
this changeset preempts failures that would have occurred on
CI testing after the new release of gr1c.