API: replace arg `option` with kwarg `solver` for func `synth.synthesize`
A default solver is now available in pure Python with any
installation of `tulip`. So the positional argument `option`
can become a keyword argument. The primary arguments that
define a synthesis problem are `specs, env, sys`.

A new user reading "solver" can more easily guess what the
argument configures. This is why it is used instead of
"option", which is more general. This naming does not
preclude using it the same argument to pass finer options.