Write `exec` in the syntax of a built-in function
because it is a built-in function in Python 3, whereas it is a simple
statement in Python 2.7 that can be written in a tuple form.

Details for the case of Python 2.7 are documented at
https://docs.python.org/2.7/reference/simple_stmts.html#the-exec-statement

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.