Use division operator with semantics of Python 3
In cases where floor division was implicitly used, now `//` is used
explicitly. E.g., observe changes to random_world() in tulip.gridworld.

For further details regarding the changed semantics, read
PEP 238, Changing the Division Operator,
https://www.python.org/dev/peps/pep-0238/

In several cases, `from __future__ import division` is added according
to the recommendation of the `futurize` command-line program, which is
part of the `python-future` package (http://www.python-future.org/).

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.