Replace dict iter methods
The dict methods `iteritems`, `iterkeys`, and `itervalues` no longer
exist in Python 3.

This change causes different behavior in Python 2.7 vs. Python 3. In
Python 2.7, `dict().items()` is a list, whereas in Python3 it is a
generator.  Though we can use generators across versions, the
implementation is simpler if we favor Python 3 in this way. The same
comment applies for `keys` and `values`. Discussion on this topic is
in issue #44.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.