Use new-style classes everywhere
Change classes to inherit from `object` when they would have otherwise
been old-style (or "classic") classes in Python 2.7. This distinction
does not exist in Python 3, i.e., moving forward, there are only
"new-style" classes.

At the time of this commit, I am not aware that old-style class
semantics were being distinctly used in TuLiP. Nonetheless, it is
better to be explicit about expectations and, thus, inherit from
`object` while we continue to support Python 2.7.

Further reading on this topic can be found at
https://docs.python.org/2.7/reference/datamodel.html#new-style-and-classic-classes
https://www.python.org/doc/newstyle/

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.