Use list() or list comprehensions where needed
In Python 3, the built-in functions `map`, `filter`, `zip`, and
`range` provide iterators, whereas in Python 2.7 they return instances
of `list`. Similarly, the methods `keys`, `items`, and `values` of
`dict` now provide views. An announcement about these changes appears
in the "What's New" section of the Python documentation,
https://docs.python.org/3.5/whatsnew/3.0.html#views-and-iterators-instead-of-lists

This changeset modifies usage of these affected functions where an
iterator return value cannot be handled. In some cases, broader design
changes are feasible to allow iterators, but such changes would be
outside the scope of the current effort to target Python 3.

Most cases of source code that might need to be changed I found using
the `futurize` command-line program, which is part of the
`python-future` package (http://www.python-future.org/). It provided
many false positives.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.