Use `universal_newlines=True` in subprocess.Popen
In Python 3, text and binary data streams are treated distinctly from
each other, in particular compared to how they were handled in Python 2.7.
An announcement about these changes appears in the "What's New" section of
the Python documentation,
https://docs.python.org/3.5/whatsnew/3.0.html#text-vs-data-instead-of-unicode-vs-8-bit

One consequence is that (binary) data streams can no longer silently
appear where text strings are expected when reading and writing files.
For subprocess communication through file-like interfaces to stdin,
stdout, and stderr, this changeset adds the keyword parameter
`universal_newlines=True` to calls of the constructor method of
subprocess.Popen or convenience functions.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.