Delete reference to private (hidden) type
Several functions in tulip.interfaces.gr1c for working with XML
elements type-checked given variables against a private (hidden) class
from previous versions of xml.etree.ElementTree. As expected,
performing such a check is fragile. Now the code is broken, so in this
changeset, the protective check is removed in favor of trait-oriented
programming (also known as duck-typing).