Handle `bytes` in Python 3 vs. 2 using try-except
To support using binary mode in `TemporaryFile` across Python versions 3
and 2.7, this changeset introduces a try-except clause that first
attempts the the appropriate `bytes` call for Python 3, where an
encoding is explicitly given. If that fails, then the case of Python 2.7
is assumed.

Note the recommendation to open temporary files in binary mode from the
documentation of `tempfile` in the Python standard library, as available
at <https://docs.python.org/3.5/library/tempfile.html#tempfile.TemporaryFile>:
"Binary mode is used so that it behaves consistently on all platforms
without regard for the data that is stored."

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.