Replace `raw_input` with equiv. `stdin.readline`
In Python 3, `raw_input` is renamed to `input`. Instead of creating a
wrapped `input` that backports this behavior to Python 2.7, this
changeset replaces `raw_input` with an explicit print() of the message
followed by stdin.readline().rstrip('\r\n'). The usage here of
`stdin.readline` extends an example from PEP 3111,
https://www.python.org/dev/peps/pep-3111/

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.