Wrap dict.values() return value in iter()
Previously the code modified by this changeset had the type form
`dict().itervalues()`, which can be given to the built-in function
next(). As part of the transition to Python 3, `itervalues` was
replaced with `values`, so this changeset uses `iter()` to obtain an
iterable appropriate for `next()`.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.