Use io.StringIO instead of cStringIO.StringIO
The StringIO and cStringIO modules no longer exist in Python 3. The
new module to use instead is `io`.

I did not perform experiments to learn whether there is a significant
increase of runtimes in the affected code of transys.export.graph2dot
when using the class StringIO from `io` in Python 2.7 instead of from
`cStringIO`. However, there is no obvious change in time to complete
the full suite of tests on my local Python 2.7 platform.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.