Use function-like semantics of exec in Python 2.7
Before applying this changeset, in Python 2.7, the generated code from
tulip.dumpsmach is executed in the current scope of `exec` [1], which
is possible because `exec` is a statement, not a function. However,
syntax allows function-like behavior of `exec` [1]. This changeset
uses this syntax to maintain support for Python 2.7 while realizing
the change to Python 3 where `exec` is a built-in function [2].

[1] https://docs.python.org/2.7/reference/simple_stmts.html#the-exec-statement
[2] https://docs.python.org/3.5/library/functions.html#exec

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.