Replace `isinstance` of basestring with str
Entirely similar to commit d6ecf36d5dd8878b27e505c52c005c446a1de528.
Consult that commit message for details. In summary, `basestring` is
not in Python 3; type-checking for it is equivalent to checking for
`str` or `unicode`, but `unicode` is not Python 3. From my
understanding of affected code in transys/export/graph2dot.py, only
checking for type `str` is sufficient.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.