BUG: Fix misprint of variable name `constraint`
Spelling error in declaration manifested as

    UnboundLocalError: local variable 'constraint' referenced before assignment

I noticed this error while running the base suite of tests using
Python 3.5.3. Surprisingly, the error does not occur when running the
full suite (of which the base suite is a subset) using Python 2.7.13.
In all cases the OS is Fedora 25 GNU/Linux on x86_64.