Enumerate `filter` return value in `mutex`
As described in the message of commit 2f4f67e46e1edd160db268f8abd320993247fd87,
in Python 3 `filter` returns an iterator. To recover the behavior as
expected from Python 2.7, the value obtained from the call to `filter`
in the function `mutex` of tulip/synth.py is wrapped in list().

In this changeset, two conditional branches are combined into one to
improve code clarity.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.