Define __hash__ in class Terminal in ast.py
as discussed in issue #44
(https://github.com/tulip-control/tulip-control/issues/44).
In summary, default values for the special method __hash__ are
different in Python 3 and 2.7 in a way that affects the implementation
of the class Terminal.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.