Expand hybrid modes iterator into `list`
The `modes` property of the class SwitchedSysDyn in tulip.hybrid has
the type of dict.keys(), unless `dynamics` of the object is None.
In this changeset, two cases of indexing into `modes` are preceded by
expanding it into a local `list` object, which recovers correct
operation in Python 3.

This changeset is part of a sequence that transitions TuLiP to Python 3
while maintaining support for Python 2.7.