MAI: use division operator with Python 3 semantics
Following [1]. With the values at [2] (which are integers) for
the variables `prob` and `genfail`, the operator `/` resulted
in floor division in Python 2.7. So applying `int` would be unnecessary.

However, the surrounding `int`, together with the observation
that `prob` represents a probability suggest that floats are
intended to be allowed as values.

Assuming so, this change modifies the behavior of `prob / genfail`
for the case of integers only, but the expression `int(prob / genfail)`
gives the same result, because `int` truncates towards zero.

[1] 22170ca462f549bec617b21901ae44639f1b9f1b
[2] https://github.com/tulip-control/tulip-control/blob/master/contrib/AES/AES_specgen.py#L1000