CONTRIB: generic bootstrapping of VMs to have TuLiP
tulip-in-a-box.py is a small Python script that uses the Travis CI
configuration in .travis.yml to install and fully test TuLiP on a
Debian-like distribution.

tulip-in-a-box.py requires PyYAML and a local copy of the TuLiP
sourcetree which is to be installed and tested. To further automate
this process, vm-bootstrap.sh is a shell script that will install Git,
perform the clone, etc., and finally run tulip-in-a-box.py.

Both vm-bootstrap.sh and tulip-in-a-box.py have been demonstrated on a
Debian 9 (stretch) GNU/Linux instance in Google Cloud Platform
(https://cloud.google.com/).