Avoid disposal of initial transitions
I proposed this change so that the transition from the 'Sinit' state to initial nodes are considered in the exportation as well. The previous code recreate transitions from successors node of 'Sinit' (already considered in the complete transition exportation) to their successors, which cause redundant transitions. In the proposed code 'Sinit' is created and transitions  from it to its successors are considered as 'init_trans'.