Errors when generating Stateflow
I was apple to generate the Matlab file and the Stateflow, but I could not run it in Stateflow because of the folowing error:

1) 
Chart 'TulipController' has unresolved symbols.
I think that the problem comes from the fact that the env. input signal has integer values (e.x. park in discrete example which is build using load_discrete.m to be selected randomly between 0 or 1) . Whereas for the generated TulipController, it takes as input and output, the values of True and False (not the boolean ones of matlab with first minuscule letters). Therefore, a change to the input and output values of TulipController should be done to either be 0s or 1s.
Other option:  change to all minuscule letters ( change of size problem).

2) 
The lines 110..113 and 164..167 belong to the switched case, so they should be surrounded by if.

3) 
Running the simulation with no unconditional transition gives:

Chart 'TulipController' has no unconditional default path to a state.
This may lead to a state inconsistency error during runtime. You can also configure the diagnostic by clicking here.

So lines 372 to 403 are added to create one.

4)
The indexes of the starting and ending states of transitions is set by finding the corresponding states name in the original 'TS.states'.
The reason for this is that naming of states in the original state-machine is not always sequential, whereas naming in state_handles is.  

5) 
Finally, the presence of 'Sinit' and initial transitions from it are considered (after considering the modifications in 'tomatlab.py' to add them to TS structure (.m file)).