REF: omit kw args with default values
The arguments `bool_states` and `bool_actions` of the functions
`sys_to_spec` and `env_to_spec` have been deprecated in [1].
The values passed at these call sites are the default values
in the signatures of those functions.

So there is no reason to pass them, considering also that
the values of those arguments will remain unchanged in the future.

[1] 90e6c923ceb420d1b9492535f34a8cee93f59582