BUG: convert strategy to `MealyMachine` in `synthesize_many`
The error here was a missing conversion from an enumerated
state graph with the synthesized transitions to a Mealy machine.
This error appears to have been introduced in
74807640a1a1fe6da2e579189dc5ce4a0b0ea118.

The conversion was previously within the module
`tulip.interfaces.gr1c`. When the conversion was moved to
`tulip.synth`, the function `synthesize` was updated,
but the function `synthesize_many` was forgotten.
This is how the error was introduced.