CI: new location for Lily
the old URL,
https://www.iaik.tugraz.at/content/research/opensource/lily/lily-2008-06-19.tar.gz
is not valid now.

we should contact the original authors eventually, but for now I
posted a copy of the tarball to the cudd-mirror SourceForge project
(https://sourceforge.net/projects/cudd-mirror/files/) that I created
when the CUDD website was removed. these are all open source, so
hosting mirrors is OK, and the SHA hashes provide certificates that
the data are as on their original websites.