API: copy only declared variables from strategy
The graph returned by the synthesis interfaces may
contain more node attributes than there are
declared variables (these interfaces are called
in the function `tulip.synth._synthesize`).

For example, the function
`tulip.interfaces.omega.synthesize_enumerated_streett`
returns a graph with the additional variable `_goal`.

This change annotates the machine transitions with
only attributes from the strategy that are declared
as variables in the specification.

These declared variables are defined as the inputs
and outputs of the machine in the function
`tulip.synth.strategy2mealy`.

Otherwise, a warning message would be raised about
the attributes that are not defined as inputs or
outputs of the machine.